Move to CROSS_COMPILE env variable
We want to be able to use native compilers as well as different architecture
compilers. The usual way (at least in U-Boot and Linux land) this works is by
defining an environment variable called CROSS_COMPILE that we rely on to point
us to the correct toolchain.

Move to the same model, so that we can compile with different compilers.

Signed-off-by: Alexander Graf <agraf@suse.de>