Use uintptr_t where we can
A lot of code today converts pointers to integers using direct
uint32_t casts. This works fine on 32bit platforms, but makes
the compiler terribly unhappy on 64bit ones.

In practice, it wouldn't matter, since all pointers on the linear
address space on the RPis are 32bit anyway, but let's be nice and
convert them to uintptr_t's instead.

This patch should result in no binary change on 32bit targets.

Signed-off-by: Alexander Graf <agraf@suse.de>