Move config file parsing into separate file
We will soon need to parse multiboot as well as config.txt style
configuration files. Move the config file helper logic into its
own file so that it's easier to share code.

To get rid of the sizeof() dependency to determine the array size
of the methods we pass, we introduce a NULL terminated finalizing
struct element.

Signed-off-by: Alexander Graf <agraf@suse.de>