Add qemufw target
QEMU is in lack of a working firmware for its Raspberry Pi model(s). The
only way to boot a RPi VM is by passing in -bios or -kernel manually.

This patch adds a new qemufw.elf target to the rpi-boot code base that
generates a simple firmware that reads config.txt and boots the kernel
described by it.

Signed-off-by: Alexander Graf